Refactor code to move functions into helper files
